Clean up helpers/ folder
