+++ src/helpers/storageHelper.js
@@ -24,7 +24,7 @@ export const uploadPictureToFirebase = (
     // Listen for state changes, errors, and completion of the upload.
     uploadTask.on(
       firebase.storage.TaskEvent.STATE_CHANGED, // or 'state_changed'
-      function(snapshot) {
+      snapshot => {
         // Get task progress, including the number of bytes uploaded and
         // the total number of bytes to be uploaded
         let progress =
@@ -44,31 +44,34 @@ export const uploadPictureToFirebase = (
             break;
         }
       },
-      function(error) {
+      error => {
         // A full list of error codes is available at
         // https://firebase.google.com/docs/storage/web/handle-errors
-        console.error(error);
         switch (error.code) {
           case 'storage/unauthorized':
             // User doesn't have permission to access the object
+            console.error('Storage unauthorized: ', error);
             break;
 
           case 'storage/canceled':
             // User canceled the upload
+            console.error('Storage cancelled: ', error);
             break;
 
           case 'storage/unknown':
             // Unknown error occurred, inspect error.serverResponse
+            console.error('Unknown error : ', error);
             break;
           default:
             break;
         }
       },
-      function() {
+      () => {
         // Upload completed successfully, now we can get the download URL
-        uploadTask.snapshot.ref.getDownloadURL().then(function(downloadURL) {
+        uploadTask.snapshot.ref.getDownloadURL().then(downloadURL => {
           if (monitor) {
-            monitor(1); // 100% uploaded
+            // 100% uploaded
+            monitor(1);
           }
           if (callback) {
             callback(downloadURL);
